+++ modules/flowable-cmmn-rest/src/test/java/org/flowable/cmmn/rest/DispatcherServletConfiguration.java
@@ -14,7 +14,7 @@ package org.flowable.cmmn.rest;
 
 import java.util.List;
 
-import org.flowable.cmmn.rest.service.api.PutAwareCommonsMultipartResolver;
+import org.flowable.common.rest.multipart.PutAwareCommonsMultipartResolver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
