+++ modules/flowable-content-rest/src/test/java/org/flowable/content/rest/DispatcherServletConfiguration.java
@@ -14,6 +14,7 @@ package org.flowable.content.rest;
 
 import java.util.List;
 
+import org.flowable.common.rest.multipart.PutAwareCommonsMultipartResolver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
