+++ modules/flowable-dmn-rest/src/test/java/org/flowable/dmn/rest/DispatcherServletConfiguration.java
@@ -16,7 +16,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.util.List;
 
-import org.flowable.dmn.rest.service.api.PutAwareCommonsMultipartResolver;
+import org.flowable.common.rest.multipart.PutAwareCommonsMultipartResolver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
