+++ modules/flowable-spring-boot/flowable-spring-boot-starters/flowable-spring-boot-autoconfigure/src/main/java/org/flowable/spring/boot/DispatcherServletConfiguration.java
@@ -14,11 +14,12 @@ package org.flowable.spring.boot;
 
 import java.util.List;
 
-import org.flowable.rest.service.api.PutAwareCommonsMultipartResolver;
+import org.flowable.common.rest.multipart.PutAwareStandardServletMultiPartResolver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.autoconfigure.web.MultipartProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
@@ -62,8 +63,9 @@ public class DispatcherServletConfiguration extends WebMvcConfigurationSupport {
     }
 
     @Bean
-    public MultipartResolver multipartResolver() {
-        PutAwareCommonsMultipartResolver multipartResolver = new PutAwareCommonsMultipartResolver();
+    public MultipartResolver multipartResolver(MultipartProperties multipartProperties) {
+        PutAwareStandardServletMultiPartResolver multipartResolver = new PutAwareStandardServletMultiPartResolver();
+        multipartResolver.setResolveLazily(multipartProperties.isResolveLazily());
         return multipartResolver;
     }
